Fix a bug introduced by parse-latin@0.1.3
The full-stop merger, introduced in wooorm/parse-latin@b794084,
tries hard to merge remaining full stops into neighbouring
words. Problem is, it doesn't known about ellipsis with spaces.
In the case where ellipsis (with spaces) occurs at the start of
a document, the last is merged into the following word when not
followed by a space.

This update fixes this problem.